feat: make dashboard great again
